#248 refactor : 알림 Panel 반응형 적용
